Refactor files for class name coding style
