+++ src/main/java/numerator/Numerator.java
@@ -2,6 +2,7 @@ package numerator;
 
 import numerator.exceptions.NumeratorException;
 import numerator.exceptions.storage.LoadingException;
+import numerator.task.TaskList;
 
 import java.nio.file.Path;
 import java.util.Scanner;
